+++ lib/utilities/manifest-utilities.js
@@ -9,7 +9,7 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
     });
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.isWorkloadEntity = exports.getUpdatedManifestFiles = exports.updateImagePullSecrets = exports.substituteImageNameInSpecFile = exports.getDeleteCmdArgs = exports.createKubectlArgs = exports.getKubectl = exports.getManifestFiles = void 0;
+exports.isWorkloadEntity = exports.getUpdatedManifestFiles = exports.updateImagePullSecrets = exports.getDeleteCmdArgs = exports.createKubectlArgs = exports.getKubectl = exports.getManifestFiles = void 0;
 const core = require("@actions/core");
 const fs = require("fs");
 const yaml = require("js-yaml");
@@ -72,61 +72,6 @@ exports.getDeleteCmdArgs = getDeleteCmdArgs;
     This substituteImageNameInSpecFile function would return
         return Value: `image: "example/example-image:identifiertag"`
 */
-function substituteImageNameInSpecFile(inputObject, containers) {
-    if (!inputObject || !inputObject.spec || !containers) {
-        return;
-    }
-    if (inputObject.spec.template && !!inputObject.spec.template.spec) {
-        if (inputObject.spec.template.spec.containers) {
-            updateContainers(inputObject.spec.template.spec.containers, containers);
-        }
-        if (inputObject.spec.template.spec.initContainers) {
-            updateContainers(inputObject.spec.template.spec.initContainers, containers);
-        }
-        return;
-    }
-    if (inputObject.spec.jobTemplate && inputObject.spec.jobTemplate.spec && inputObject.spec.jobTemplate.spec.template && inputObject.spec.jobTemplate.spec.template.spec) {
-        if (inputObject.spec.jobTemplate.spec.template.spec.containers) {
-            updateContainers(inputObject.spec.jobTemplate.spec.template.spec.containers, containers);
-        }
-        if (inputObject.spec.jobTemplate.spec.template.spec.initContainers) {
-            updateContainers(inputObject.spec.jobTemplate.spec.template.spec.initContainers, containers);
-        }
-        return;
-    }
-    if (inputObject.spec.containers) {
-        updateContainers(inputObject.spec.containers, containers);
-    }
-    if (inputObject.spec.initContainers) {
-        updateContainers(inputObject.spec.initContainers, containers);
-    }
-}
-exports.substituteImageNameInSpecFile = substituteImageNameInSpecFile;
-function updateContainers(inputContainers, images) {
-    if (!inputContainers || inputContainers.length === 0) {
-        return inputContainers;
-    }
-    inputContainers.forEach((inputContainer) => {
-        const imageName = extractImageName(inputContainer.image.trim());
-        images.forEach(image => {
-            if (extractImageName(image) === imageName) {
-                inputContainer.image = image;
-            }
-        });
-    });
-}
-function extractImageName(imageName) {
-    let img = '';
-    if (imageName.indexOf('/') > 0) {
-        const registry = imageName.substring(0, imageName.indexOf('/'));
-        const imgName = imageName.substring(imageName.indexOf('/') + 1).split(':')[0];
-        img = `${registry}/${imgName}`;
-    }
-    else {
-        img = imageName.split(':')[0];
-    }
-    return img;
-}
 function createInlineArray(str) {
     if (typeof str === 'string') {
         return str;
@@ -225,34 +170,6 @@ function substituteImageNameInSpecContent(currentString, imageName, imageNameWit
         return acc + line + '\n';
     }, '');
 }
-function updateContainerImagesInManifestFiles(filePaths, containers) {
-    if (!!containers && containers.length > 0) {
-        const newObjectsList = [];
-        const tempDirectory = fileHelper.getTempDirectory();
-        filePaths.forEach((filePath) => {
-            const fileContents = fs.readFileSync(filePath).toString();
-            yaml.safeLoadAll(fileContents, function (inputObject) {
-                if (!!inputObject && !!inputObject.kind) {
-                    const kind = inputObject.kind;
-                    if (KubernetesObjectUtility.isWorkloadEntity(kind)) {
-                        substituteImageNameInSpecFile(inputObject, containers);
-                    }
-                    else if (utility_1.isEqual(kind, 'list', true)) {
-                        let items = inputObject.items;
-                        if (items.length > 0) {
-                            items.forEach((item) => substituteImageNameInSpecFile(item, containers));
-                        }
-                    }
-                    newObjectsList.push(inputObject);
-                }
-            });
-        });
-        core.debug('New K8s objects after updating container images are :' + JSON.stringify(newObjectsList));
-        const newFilePaths = fileHelper.writeObjectsToFile(newObjectsList);
-        return newFilePaths;
-    }
-    return filePaths;
-}
 function updateImagePullSecrets(inputObject, newImagePullSecrets) {
     if (!inputObject || !inputObject.spec || !newImagePullSecrets) {
         return;
@@ -272,22 +189,40 @@ function updateImagePullSecrets(inputObject, newImagePullSecrets) {
     setImagePullSecrets(inputObject, existingImagePullSecretObjects);
 }
 exports.updateImagePullSecrets = updateImagePullSecrets;
-function updateImagePullSecretsInManifestFiles(filePaths, imagePullSecrets) {
-    if (!!imagePullSecrets && imagePullSecrets.length > 0) {
-        const newObjectsList = [];
+function updateResourceObjects(filePaths, imagePullSecrets, containers) {
+    if ((!!imagePullSecrets && imagePullSecrets.length > 0) || (!!containers && containers.length > 0)) {
+        let newObjectsList = [];
         filePaths.forEach((filePath) => {
             const fileContents = fs.readFileSync(filePath).toString();
             yaml.safeLoadAll(fileContents, function (inputObject) {
                 if (!!inputObject && !!inputObject.kind) {
                     const kind = inputObject.kind;
                     if (KubernetesObjectUtility.isWorkloadEntity(kind)) {
-                        KubernetesObjectUtility.updateImagePullSecrets(inputObject, imagePullSecrets, false);
+                        if (!!imagePullSecrets && imagePullSecrets.length > 0) {
+                            KubernetesObjectUtility.updateImagePullSecrets(inputObject, imagePullSecrets, false);
+                        }
+                        if (!!containers && containers.length > 0) {
+                            KubernetesObjectUtility.substituteImageNameInSpecFile(inputObject, containers);
+                        }
+                    }
+                    else if (utility_1.isEqual(kind, 'list', true)) {
+                        let items = inputObject.items;
+                        if (items.length > 0) {
+                            items.forEach((item) => {
+                                if (!!imagePullSecrets && imagePullSecrets.length > 0) {
+                                    KubernetesObjectUtility.updateImagePullSecrets(item, imagePullSecrets, false);
+                                }
+                                if (!!containers && containers.length > 0) {
+                                    KubernetesObjectUtility.substituteImageNameInSpecFile(item, containers);
+                                }
+                            });
+                        }
                     }
                     newObjectsList.push(inputObject);
                 }
             });
         });
-        core.debug('New K8s objects after adding imagePullSecrets are :' + JSON.stringify(newObjectsList));
+        core.debug('New K8s objects after adding imagePullSecrets and updating container images are :' + JSON.stringify(newObjectsList));
         const newFilePaths = fileHelper.writeObjectsToFile(newObjectsList);
         return newFilePaths;
     }
@@ -298,10 +233,7 @@ function getUpdatedManifestFiles(manifestFilePaths) {
     if (!inputManifestFiles || inputManifestFiles.length === 0) {
         throw new Error(`ManifestFileNotFound : ${manifestFilePaths}`);
     }
-    // artifact substitution
-    inputManifestFiles = updateContainerImagesInManifestFiles(inputManifestFiles, TaskInputParameters.containers);
-    // imagePullSecrets addition
-    inputManifestFiles = updateImagePullSecretsInManifestFiles(inputManifestFiles, TaskInputParameters.imagePullSecrets);
+    inputManifestFiles = updateResourceObjects(inputManifestFiles, TaskInputParameters.imagePullSecrets, TaskInputParameters.containers);
     return inputManifestFiles;
 }
 exports.getUpdatedManifestFiles = getUpdatedManifestFiles;
