+++ lib/utilities/resource-object-utility.js
@@ -1,6 +1,6 @@
 'use strict';
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.getResources = exports.updateSelectorLabels = exports.updateSpecLabels = exports.updateImagePullSecrets = exports.updateObjectAnnotations = exports.updateObjectLabels = exports.getReplicaCount = exports.isIngressEntity = exports.isServiceEntity = exports.isWorkloadEntity = exports.isDeploymentEntity = void 0;
+exports.substituteImageNameInSpecFile = exports.getResources = exports.updateSelectorLabels = exports.updateSpecLabels = exports.updateImagePullSecrets = exports.updateObjectAnnotations = exports.updateObjectLabels = exports.getReplicaCount = exports.isIngressEntity = exports.isServiceEntity = exports.isWorkloadEntity = exports.isDeploymentEntity = void 0;
 const fs = require("fs");
 const core = require("@actions/core");
 const yaml = require("js-yaml");
@@ -195,6 +195,61 @@ function getResources(filePaths, filterResourceTypes) {
     return resources;
 }
 exports.getResources = getResources;
+function substituteImageNameInSpecFile(inputObject, containers) {
+    if (!inputObject || !inputObject.spec || !containers) {
+        return;
+    }
+    if (inputObject.spec.template && !!inputObject.spec.template.spec) {
+        if (inputObject.spec.template.spec.containers) {
+            updateContainers(inputObject.spec.template.spec.containers, containers);
+        }
+        if (inputObject.spec.template.spec.initContainers) {
+            updateContainers(inputObject.spec.template.spec.initContainers, containers);
+        }
+        return;
+    }
+    if (inputObject.spec.jobTemplate && inputObject.spec.jobTemplate.spec && inputObject.spec.jobTemplate.spec.template && inputObject.spec.jobTemplate.spec.template.spec) {
+        if (inputObject.spec.jobTemplate.spec.template.spec.containers) {
+            updateContainers(inputObject.spec.jobTemplate.spec.template.spec.containers, containers);
+        }
+        if (inputObject.spec.jobTemplate.spec.template.spec.initContainers) {
+            updateContainers(inputObject.spec.jobTemplate.spec.template.spec.initContainers, containers);
+        }
+        return;
+    }
+    if (inputObject.spec.containers) {
+        updateContainers(inputObject.spec.containers, containers);
+    }
+    if (inputObject.spec.initContainers) {
+        updateContainers(inputObject.spec.initContainers, containers);
+    }
+}
+exports.substituteImageNameInSpecFile = substituteImageNameInSpecFile;
+function updateContainers(inputContainers, images) {
+    if (!inputContainers || inputContainers.length === 0) {
+        return inputContainers;
+    }
+    inputContainers.forEach((inputContainer) => {
+        const imageName = extractImageName(inputContainer.image.trim());
+        images.forEach(image => {
+            if (extractImageName(image) === imageName) {
+                inputContainer.image = image;
+            }
+        });
+    });
+}
+function extractImageName(imageName) {
+    let img = '';
+    if (imageName.indexOf('/') > 0) {
+        const registry = imageName.substring(0, imageName.indexOf('/'));
+        const imgName = imageName.substring(imageName.indexOf('/') + 1).split(':')[0];
+        img = `${registry}/${imgName}`;
+    }
+    else {
+        img = imageName.split(':')[0];
+    }
+    return img;
+}
 function getSpecLabels(inputObject) {
     if (!inputObject) {
         return null;
